fix(bf16): cast to fp32 at every linalg/serialization boundary (BUG-5)
PyTorch has no bf16 kernels for qr/geqrf, eigvalsh/eigh, svd_lowrank,
matrix_exp or fft, and bf16 tensors have no numpy dtype. Cast to fp32 at
each boundary and cast tensors back to the working dtype afterward, mirroring
the pattern already used in gue_loss.py.

Sites fixed:
  - dual_link.send_krylov_basis: serialize the SVD basis as fp32
  - bridge._randomized_power_iteration: fp32 QR, cast Q back per step
  - bridge.spectral_gap: fp32 eigvalsh
  - bridge.reorthogonalize: fp32 QR for eigenvectors and LoRA A
  - virtual_layer13._entropy (svdvals) and _hash_psi (numpy)
  - chronos_lock.unitary_wait_spin: fp32 matrix_exp, cast back to h.dtype
  - semantic_lock.anchor_hash: fp32 numpy serialization
  - ghost_layer.encode_shard: fp32 FFT

flux/horizons/casimir/unitary_regulator already operate in fp32/fp64 by
construction. Adds tests/test_bf16_paths.py exercising every fixed function
with CPU bf16 tensors (reproduces all four verified crash classes).

+++ unitarity_labs/core/dual_link.py
@@ -142,6 +142,8 @@ class DualNodeEntanglementBridge:
 
     def send_krylov_basis(self, krylov_basis: torch.Tensor) -> None:
         """Compress and transmit Krylov subspace via SVD low-rank."""
+        # fp32 at the linalg/serialization boundary: bf16/fp8 have no
+        # svd_lowrank kernel and no numpy dtype. Always serialize fp32.
         A = krylov_basis.float()
         # Clamp to matrix dims: small matrices can't be compressed to a larger rank
         q_eff = min(self.krylov_dim, A.shape[-2], A.shape[-1])
@@ -151,7 +153,7 @@ class DualNodeEntanglementBridge:
             U, _, _ = torch.svd_lowrank(A, q=q_eff)
         msg = {
             'node': self.node_id,
-            'basis': U.cpu().numpy(),
+            'basis': U.to(torch.float32).cpu().numpy(),
             'timestamp': time.monotonic(),
         }
         self.pub.send_pyobj(msg)
